feat: #811 Add tracebility comments in Effective Set
